% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_ar_psd}
\alias{compute_gazepoint_pyhrv_ar_psd}
\title{Compute pyHRV-style autoregressive PSD}
\usage{
compute_gazepoint_pyhrv_ar_psd(
  nni_ms,
  time_s = NULL,
  resample_hz = 4,
  order = NULL
)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{time_s}{Optional time vector in seconds.}

\item{resample_hz}{Resampling frequency.}

\item{order}{Optional AR order.}
}
\value{
List with PSD and measures.
}
\description{
Compute pyHRV-style autoregressive PSD
}
