% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_dfa}
\alias{compute_gazepoint_pyhrv_dfa}
\title{Compute detrended fluctuation analysis}
\usage{
compute_gazepoint_pyhrv_dfa(
  nni_ms,
  scales = unique(round(exp(seq(log(4), log(64), length.out = 12))))
)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{scales}{Window sizes in beats.}
}
\value{
Data frame with DFA alpha estimates.
}
\description{
Compute detrended fluctuation analysis
}
