% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_heart_rate}
\alias{compute_gazepoint_pyhrv_heart_rate}
\title{Compute heart rate from NN intervals}
\usage{
compute_gazepoint_pyhrv_heart_rate(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
Numeric heart rate in beats per minute.
}
\description{
Compute heart rate from NN intervals
}
