% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_hr_parameters}
\alias{compute_gazepoint_pyhrv_hr_parameters}
\title{Compute pyHRV-style heart-rate parameters}
\usage{
compute_gazepoint_pyhrv_hr_parameters(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
Data frame of heart-rate parameters.
}
\description{
Compute pyHRV-style heart-rate parameters
}
