% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_lomb_psd}
\alias{compute_gazepoint_pyhrv_lomb_psd}
\title{Compute pyHRV-style Lomb PSD}
\usage{
compute_gazepoint_pyhrv_lomb_psd(
  nni_ms,
  time_s = NULL,
  min_hz = 0.003,
  max_hz = 0.4,
  n_freq = 512
)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{time_s}{Optional time vector in seconds.}

\item{min_hz}{Minimum frequency.}

\item{max_hz}{Maximum frequency.}

\item{n_freq}{Number of frequencies.}
}
\value{
List with PSD and measures.
}
\description{
Compute pyHRV-style Lomb PSD
}
