% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_nn20}
\alias{compute_gazepoint_pyhrv_nn20}
\title{Compute NN20 and pNN20}
\usage{
compute_gazepoint_pyhrv_nn20(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
Data frame with NN20 and pNN20.
}
\description{
Compute NN20 and pNN20
}
