% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_nn50}
\alias{compute_gazepoint_pyhrv_nn50}
\title{Compute NN50 and pNN50}
\usage{
compute_gazepoint_pyhrv_nn50(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
Data frame with NN50 and pNN50.
}
\description{
Compute NN50 and pNN50
}
