% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_nn_diff}
\alias{compute_gazepoint_pyhrv_nn_diff}
\title{Compute NN interval differences}
\usage{
compute_gazepoint_pyhrv_nn_diff(nni_ms, absolute = FALSE)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{absolute}{If TRUE, return absolute differences.}
}
\value{
Numeric successive NN differences in milliseconds.
}
\description{
Compute NN interval differences
}
