% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_nni_differences_parameters}
\alias{compute_gazepoint_pyhrv_nni_differences_parameters}
\title{Compute pyHRV-style NNI difference parameters}
\usage{
compute_gazepoint_pyhrv_nni_differences_parameters(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
Data frame of successive-difference parameters.
}
\description{
Compute pyHRV-style NNI difference parameters
}
