% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_nnxx}
\alias{compute_gazepoint_pyhrv_nnxx}
\title{Compute NNxx and pNNxx}
\usage{
compute_gazepoint_pyhrv_nnxx(nni_ms, threshold_ms = 50)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{threshold_ms}{Threshold in milliseconds.}
}
\value{
Data frame with NNxx and pNNxx.
}
\description{
Compute NNxx and pNNxx
}
