% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_nonlinear}
\alias{compute_gazepoint_pyhrv_nonlinear}
\title{Compute pyHRV-style nonlinear summary}
\usage{
compute_gazepoint_pyhrv_nonlinear(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
Data frame with nonlinear measures.
}
\description{
Compute pyHRV-style nonlinear summary
}
