% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_poincare}
\alias{compute_gazepoint_pyhrv_poincare}
\title{Compute pyHRV-style Poincare measures}
\usage{
compute_gazepoint_pyhrv_poincare(nni_ms, plot = FALSE)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{plot}{If TRUE, draw Poincare plot.}
}
\value{
Data frame of Poincare measures.
}
\description{
Compute pyHRV-style Poincare measures
}
