% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_rmssd}
\alias{compute_gazepoint_pyhrv_rmssd}
\title{Compute RMSSD}
\usage{
compute_gazepoint_pyhrv_rmssd(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
RMSSD in milliseconds.
}
\description{
Compute RMSSD
}
