% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_sample_entropy}
\alias{compute_gazepoint_pyhrv_sample_entropy}
\title{Compute sample entropy}
\usage{
compute_gazepoint_pyhrv_sample_entropy(nni_ms, m = 2L, r = NULL)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{m}{Embedding dimension.}

\item{r}{Tolerance. If NULL, 0.2 * SD is used.}
}
\value{
Sample entropy.
}
\description{
Compute sample entropy
}
