% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_sdann}
\alias{compute_gazepoint_pyhrv_sdann}
\title{Compute SDANN}
\usage{
compute_gazepoint_pyhrv_sdann(nni_ms, segment_seconds = 300)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{segment_seconds}{Segment length.}
}
\value{
SDANN in milliseconds.
}
\description{
Compute SDANN
}
