% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_sdnn}
\alias{compute_gazepoint_pyhrv_sdnn}
\title{Compute SDNN}
\usage{
compute_gazepoint_pyhrv_sdnn(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
SDNN in milliseconds.
}
\description{
Compute SDNN
}
