% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_sdnn_index}
\alias{compute_gazepoint_pyhrv_sdnn_index}
\title{Compute SDNN index}
\usage{
compute_gazepoint_pyhrv_sdnn_index(nni_ms, segment_seconds = 300)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{segment_seconds}{Segment length.}
}
\value{
SDNN index in milliseconds.
}
\description{
Compute SDNN index
}
