% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_sdsd}
\alias{compute_gazepoint_pyhrv_sdsd}
\title{Compute SDSD}
\usage{
compute_gazepoint_pyhrv_sdsd(nni_ms)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}
}
\value{
SDSD in milliseconds.
}
\description{
Compute SDSD
}
