% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_tinn}
\alias{compute_gazepoint_pyhrv_tinn}
\title{Compute TINN}
\usage{
compute_gazepoint_pyhrv_tinn(nni_ms, bin_width_ms = 7.8125)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{bin_width_ms}{Histogram bin width.}
}
\value{
Approximate TINN in milliseconds.
}
\description{
Compute TINN
}
