% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_triangular_index}
\alias{compute_gazepoint_pyhrv_triangular_index}
\title{Compute triangular index}
\usage{
compute_gazepoint_pyhrv_triangular_index(nni_ms, bin_width_ms = 7.8125)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{bin_width_ms}{Histogram bin width.}
}
\value{
HRV triangular index.
}
\description{
Compute triangular index
}
