% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{compute_gazepoint_pyhrv_welch_psd}
\alias{compute_gazepoint_pyhrv_welch_psd}
\title{Compute pyHRV-style Welch PSD}
\usage{
compute_gazepoint_pyhrv_welch_psd(
  nni_ms,
  time_s = NULL,
  resample_hz = 4,
  window_seconds = 256,
  overlap = 0.5
)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{time_s}{Optional time vector in seconds.}

\item{resample_hz}{Resampling frequency.}

\item{window_seconds}{Welch window length.}

\item{overlap}{Window overlap proportion.}
}
\value{
List with PSD and measures.
}
\description{
Compute pyHRV-style Welch PSD
}
