% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosppy-style.R
\name{compute_gazepoint_signal_band_power}
\alias{compute_gazepoint_signal_band_power}
\title{Compute signal band power}
\usage{
compute_gazepoint_signal_band_power(
  x,
  sampling_rate_hz = NULL,
  bands = list(very_low = c(0.003, 0.04), low = c(0.04, 0.15), high = c(0.15, 0.4)),
  relative = TRUE
)
}
\arguments{
\item{x}{Numeric signal or power-spectrum data frame.}

\item{sampling_rate_hz}{Sampling rate in Hz when x is a signal.}

\item{bands}{Named list of frequency bands.}

\item{relative}{If TRUE, include relative band power.}
}
\value{
Data frame of band powers.
}
\description{
Compute signal band power
}
