% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosppy-style.R
\name{compute_gazepoint_signal_power_spectrum}
\alias{compute_gazepoint_signal_power_spectrum}
\title{Compute signal power spectrum}
\usage{
compute_gazepoint_signal_power_spectrum(x, sampling_rate_hz, detrend = TRUE)
}
\arguments{
\item{x}{Numeric signal.}

\item{sampling_rate_hz}{Sampling rate in Hz.}

\item{detrend}{If TRUE, remove the mean before FFT.}
}
\value{
Data frame with frequency and power.
}
\description{
Compute signal power spectrum
}
