% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{create_gazepoint_pyhrv_time_vector}
\alias{create_gazepoint_pyhrv_time_vector}
\title{Create a time vector from NN intervals}
\usage{
create_gazepoint_pyhrv_time_vector(nni_ms, start_s = 0)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{start_s}{Start time in seconds.}
}
\value{
Numeric time vector in seconds.
}
\description{
Create a time vector from NN intervals
}
