% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartpy-extended.R
\name{estimate_gazepoint_samplerate_mstimer}
\alias{estimate_gazepoint_samplerate_mstimer}
\title{Estimate sampling rate from a millisecond timer}
\usage{
estimate_gazepoint_samplerate_mstimer(mstimer, robust = TRUE)
}
\arguments{
\item{mstimer}{Numeric millisecond timer.}

\item{robust}{If TRUE, use the median interval.}
}
\value{
Estimated sampling-rate information.
}
\description{
Estimate sampling rate from a millisecond timer
}
