% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{export_gazepoint_pyhrv_results}
\alias{export_gazepoint_pyhrv_results}
\title{Export pyHRV-style results}
\usage{
export_gazepoint_pyhrv_results(results, path)
}
\arguments{
\item{results}{Result object.}

\item{path}{Output path. Use .json or .rds.}
}
\value{
Output path invisibly.
}
\description{
Export pyHRV-style results
}
