% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartpy-extended.R
\name{filter_gazepoint_ppg_signal}
\alias{filter_gazepoint_ppg_signal}
\title{Apply generic filtering to Gazepoint pulse/PPG data}
\usage{
filter_gazepoint_ppg_signal(
  x,
  sampling_rate_hz,
  type = c("lowpass", "highpass", "bandpass", "notch"),
  low_hz = NULL,
  high_hz = NULL,
  passes = 1L
)
}
\arguments{
\item{x}{Numeric signal.}

\item{sampling_rate_hz}{Sampling rate in Hz.}

\item{type}{Filter type: lowpass, highpass, bandpass, or notch.}

\item{low_hz}{Low cutoff for highpass, bandpass, or notch.}

\item{high_hz}{High cutoff for lowpass, bandpass, or notch.}

\item{passes}{Repeated filter passes.}
}
\value{
Filtered numeric signal.
}
\description{
Apply generic filtering to Gazepoint pulse/PPG data
}
