% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartpy-extended.R
\name{flip_gazepoint_ppg_signal}
\alias{flip_gazepoint_ppg_signal}
\title{Flip a Gazepoint pulse/PPG signal}
\usage{
flip_gazepoint_ppg_signal(x, method = c("negative", "max_minus"))
}
\arguments{
\item{x}{Numeric signal.}

\item{method}{Flip method: negative or max_minus.}
}
\value{
Flipped numeric signal.
}
\description{
Flip a Gazepoint pulse/PPG signal
}
