% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-contracts.R
\name{get_gazepoint_plot_data}
\alias{get_gazepoint_plot_data}
\title{Extract stored plot data}
\usage{
get_gazepoint_plot_data(plot)
}
\arguments{
\item{plot}{A plot object returned by a gpbiometrics plotting helper.}
}
\value{
A data frame.
}
\description{
Extract stored plot data
}
