% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-contracts.R
\name{get_gazepoint_plot_settings}
\alias{get_gazepoint_plot_settings}
\title{Extract stored plot settings}
\usage{
get_gazepoint_plot_settings(plot)
}
\arguments{
\item{plot}{A plot object returned by a gpbiometrics plotting helper.}
}
\value{
A list.
}
\description{
Extract stored plot settings
}
