% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{import_gazepoint_pyhrv_results}
\alias{import_gazepoint_pyhrv_results}
\title{Import pyHRV-style results}
\usage{
import_gazepoint_pyhrv_results(path)
}
\arguments{
\item{path}{Input path. Use .json or .rds.}
}
\value{
Imported object.
}
\description{
Import pyHRV-style results
}
