% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm-style.R
\name{merge_gazepoint_recordings_pspm_style}
\alias{merge_gazepoint_recordings_pspm_style}
\title{Merge multiple Gazepoint recordings PsPM-style}
\usage{
merge_gazepoint_recordings_pspm_style(
  recordings,
  time_col = NULL,
  gap_seconds = 1,
  recording_col = "pspm_recording",
  reset_first_time = TRUE
)
}
\arguments{
\item{recordings}{List of data frames.}

\item{time_col}{Time column.}

\item{gap_seconds}{Gap inserted between recordings.}

\item{recording_col}{Output recording-id column.}

\item{reset_first_time}{If TRUE, each input time starts from zero before offsetting.}
}
\value{
Merged data frame.
}
\description{
Merge multiple Gazepoint recordings PsPM-style
}
