% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartpy-extended.R
\name{plot_gazepoint_ppg_segmentwise}
\alias{plot_gazepoint_ppg_segmentwise}
\title{Plot segmentwise Gazepoint pulse/PPG measures}
\usage{
plot_gazepoint_ppg_segmentwise(segmentwise, measure = "bpm")
}
\arguments{
\item{segmentwise}{Object returned by process_gazepoint_ppg_segmentwise().}

\item{measure}{Measure column to plot.}
}
\value{
Invisibly returns plotted data.
}
\description{
Plot segmentwise Gazepoint pulse/PPG measures
}
