% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{plot_gazepoint_pyhrv_hr_heatplot}
\alias{plot_gazepoint_pyhrv_hr_heatplot}
\title{Plot pyHRV-style heart-rate heatplot}
\usage{
plot_gazepoint_pyhrv_hr_heatplot(nni_ms, time_bins = 20, hr_bins = 20)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{time_bins}{Number of time bins.}

\item{hr_bins}{Number of heart-rate bins.}
}
\value{
Invisibly returns heatplot data.
}
\description{
Plot pyHRV-style heart-rate heatplot
}
