% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{plot_gazepoint_pyhrv_radar_chart}
\alias{plot_gazepoint_pyhrv_radar_chart}
\title{Plot pyHRV-style radar chart}
\usage{
plot_gazepoint_pyhrv_radar_chart(
  measures,
  columns = c("sdnn", "rmssd", "sdsd", "pnn50", "lf_norm", "hf_norm", "sd1", "sd2")
)
}
\arguments{
\item{measures}{Named numeric vector or one-row data frame.}

\item{columns}{Optional columns to plot.}
}
\value{
Invisibly returns plotted values.
}
\description{
Plot pyHRV-style radar chart
}
