% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyhrv-style.R
\name{plot_gazepoint_pyhrv_tachogram}
\alias{plot_gazepoint_pyhrv_tachogram}
\title{Plot pyHRV-style tachogram}
\usage{
plot_gazepoint_pyhrv_tachogram(nni_ms, time_s = NULL)
}
\arguments{
\item{nni_ms}{Numeric NN intervals in milliseconds.}

\item{time_s}{Optional time vector.}
}
\value{
Invisibly returns plotted data.
}
\description{
Plot pyHRV-style tachogram
}
