% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-permutation-advanced-guardrails.R
\name{run_gazepoint_cluster_permutation_lmer}
\alias{run_gazepoint_cluster_permutation_lmer}
\title{Guardrail for mixed-model cluster permutation}
\usage{
run_gazepoint_cluster_permutation_lmer(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
This function always errors.
}
\description{
This function is intentionally not implemented as a runnable inferential
engine.
}
