% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-permutation-advanced-guardrails.R
\name{run_gazepoint_multidimensional_cluster_permutation}
\alias{run_gazepoint_multidimensional_cluster_permutation}
\title{Guardrail for multidimensional cluster permutation}
\usage{
run_gazepoint_multidimensional_cluster_permutation(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
This function always errors.
}
\description{
This function is intentionally not implemented as a runnable inferential
engine.
}
