% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-permutation-advanced-guardrails.R
\name{run_gazepoint_tfce}
\alias{run_gazepoint_tfce}
\title{Guardrail for threshold-free cluster enhancement}
\usage{
run_gazepoint_tfce(...)
}
\arguments{
\item{...}{Ignored.}
}
\value{
This function always errors.
}
\description{
This function is intentionally not implemented as a runnable inferential
engine.
}
