% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartpy-extended.R
\name{scale_gazepoint_ppg_signal}
\alias{scale_gazepoint_ppg_signal}
\title{Scale a Gazepoint pulse/PPG signal}
\usage{
scale_gazepoint_ppg_signal(
  x,
  method = c("zscore", "minmax", "robust", "center", "none"),
  range = c(0, 1)
)
}
\arguments{
\item{x}{Numeric signal.}

\item{method}{Scaling method: zscore, minmax, robust, center, or none.}

\item{range}{Output range for minmax scaling.}
}
\value{
Scaled numeric vector.
}
\description{
Scale a Gazepoint pulse/PPG signal
}
