% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartpy-extended.R
\name{smooth_gazepoint_ppg_signal}
\alias{smooth_gazepoint_ppg_signal}
\title{Smooth a Gazepoint pulse/PPG signal}
\usage{
smooth_gazepoint_ppg_signal(
  x,
  sampling_rate_hz,
  method = c("mean", "median"),
  window_seconds = 0.1
)
}
\arguments{
\item{x}{Numeric signal.}

\item{sampling_rate_hz}{Sampling rate in Hz.}

\item{method}{Smoothing method: mean or median.}

\item{window_seconds}{Smoothing window length.}
}
\value{
Smoothed numeric signal.
}
\description{
Smooth a Gazepoint pulse/PPG signal
}
