% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspm-style.R
\name{trim_gazepoint_biometrics_pspm_style}
\alias{trim_gazepoint_biometrics_pspm_style}
\title{Trim Gazepoint biometrics data PsPM-style}
\usage{
trim_gazepoint_biometrics_pspm_style(
  data,
  start_s = NULL,
  end_s = NULL,
  time_col = NULL,
  reset_time = FALSE
)
}
\arguments{
\item{data}{Gazepoint data frame.}

\item{start_s}{Start time in seconds.}

\item{end_s}{End time in seconds.}

\item{time_col}{Time column.}

\item{reset_time}{If TRUE, reset trimmed time to start at zero.}
}
\value{
Trimmed data frame.
}
\description{
Trim Gazepoint biometrics data PsPM-style
}
